#154 Possibility to change the OverlayManager in MapView.
